Move fluid mechanics tests into the pipe domain
